+++ config/config_test.go
@@ -33,7 +33,7 @@ type nopRecvConfig struct {
 }
 
 func (nc *nopRecvConfig) Validate() error {
-	if nc.ID() != NewID("nop") {
+	if nc.ID() != NewComponentID("nop") {
 		return errInvalidRecvConfig
 	}
 	return nil
@@ -44,7 +44,7 @@ type nopExpConfig struct {
 }
 
 func (nc *nopExpConfig) Validate() error {
-	if nc.ID() != NewID("nop") {
+	if nc.ID() != NewComponentID("nop") {
 		return errInvalidExpConfig
 	}
 	return nil
@@ -55,7 +55,7 @@ type nopProcConfig struct {
 }
 
 func (nc *nopProcConfig) Validate() error {
-	if nc.ID() != NewID("nop") {
+	if nc.ID() != NewComponentID("nop") {
 		return errInvalidProcConfig
 	}
 	return nil
@@ -66,7 +66,7 @@ type nopExtConfig struct {
 }
 
 func (nc *nopExtConfig) Validate() error {
-	if nc.ID() != NewID("nop") {
+	if nc.ID() != NewComponentID("nop") {
 		return errInvalidExtConfig
 	}
 	return nil
@@ -173,8 +173,8 @@ func TestConfigValidate(t *testing.T) {
 			name: "invalid-receiver-config",
 			cfgFn: func() *Config {
 				cfg := generateConfig()
-				cfg.Receivers[NewID("nop")] = &nopRecvConfig{
-					ReceiverSettings: NewReceiverSettings(NewID("invalid_rec_type")),
+				cfg.Receivers[NewComponentID("nop")] = &nopRecvConfig{
+					ReceiverSettings: NewReceiverSettings(NewComponentID("invalid_rec_type")),
 				}
 				return cfg
 			},
@@ -184,8 +184,8 @@ func TestConfigValidate(t *testing.T) {
 			name: "invalid-exporter-config",
 			cfgFn: func() *Config {
 				cfg := generateConfig()
-				cfg.Exporters[NewID("nop")] = &nopExpConfig{
-					ExporterSettings: NewExporterSettings(NewID("invalid_rec_type")),
+				cfg.Exporters[NewComponentID("nop")] = &nopExpConfig{
+					ExporterSettings: NewExporterSettings(NewComponentID("invalid_rec_type")),
 				}
 				return cfg
 			},
@@ -195,8 +195,8 @@ func TestConfigValidate(t *testing.T) {
 			name: "invalid-processor-config",
 			cfgFn: func() *Config {
 				cfg := generateConfig()
-				cfg.Processors[NewID("nop")] = &nopProcConfig{
-					ProcessorSettings: NewProcessorSettings(NewID("invalid_rec_type")),
+				cfg.Processors[NewComponentID("nop")] = &nopProcConfig{
+					ProcessorSettings: NewProcessorSettings(NewComponentID("invalid_rec_type")),
 				}
 				return cfg
 			},
@@ -206,8 +206,8 @@ func TestConfigValidate(t *testing.T) {
 			name: "invalid-extension-config",
 			cfgFn: func() *Config {
 				cfg := generateConfig()
-				cfg.Extensions[NewID("nop")] = &nopExtConfig{
-					ExtensionSettings: NewExtensionSettings(NewID("invalid_rec_type")),
+				cfg.Extensions[NewComponentID("nop")] = &nopExtConfig{
+					ExtensionSettings: NewExtensionSettings(NewComponentID("invalid_rec_type")),
 				}
 				return cfg
 			},
@@ -235,35 +235,35 @@ func TestConfigValidate(t *testing.T) {
 func generateConfig() *Config {
 	return &Config{
 		Receivers: map[ComponentID]Receiver{
-			NewID("nop"): &nopRecvConfig{
-				ReceiverSettings: NewReceiverSettings(NewID("nop")),
+			NewComponentID("nop"): &nopRecvConfig{
+				ReceiverSettings: NewReceiverSettings(NewComponentID("nop")),
 			},
 		},
 		Exporters: map[ComponentID]Exporter{
-			NewID("nop"): &nopExpConfig{
-				ExporterSettings: NewExporterSettings(NewID("nop")),
+			NewComponentID("nop"): &nopExpConfig{
+				ExporterSettings: NewExporterSettings(NewComponentID("nop")),
 			},
 		},
 		Processors: map[ComponentID]Processor{
-			NewID("nop"): &nopProcConfig{
-				ProcessorSettings: NewProcessorSettings(NewID("nop")),
+			NewComponentID("nop"): &nopProcConfig{
+				ProcessorSettings: NewProcessorSettings(NewComponentID("nop")),
 			},
 		},
 		Extensions: map[ComponentID]Extension{
-			NewID("nop"): &nopExtConfig{
-				ExtensionSettings: NewExtensionSettings(NewID("nop")),
+			NewComponentID("nop"): &nopExtConfig{
+				ExtensionSettings: NewExtensionSettings(NewComponentID("nop")),
 			},
 		},
 		Service: Service{
 			Telemetry:  ServiceTelemetry{Logs: ServiceTelemetryLogs{Level: zapcore.DebugLevel, Development: true, Encoding: "console"}},
-			Extensions: []ComponentID{NewID("nop")},
+			Extensions: []ComponentID{NewComponentID("nop")},
 			Pipelines: map[string]*Pipeline{
 				"traces": {
 					Name:       "traces",
 					InputType:  TracesDataType,
-					Receivers:  []ComponentID{NewID("nop")},
-					Processors: []ComponentID{NewID("nop")},
-					Exporters:  []ComponentID{NewID("nop")},
+					Receivers:  []ComponentID{NewComponentID("nop")},
+					Processors: []ComponentID{NewComponentID("nop")},
+					Exporters:  []ComponentID{NewComponentID("nop")},
 				},
 			},
 		},