+++ config/configunmarshaler/defaultunmarshaler.go
@@ -209,10 +209,9 @@ func unmarshalExtensions(exts map[string]map[string]interface{}, factories map[c
 	// Iterate over extensions and create a config for each.
 	for key, value := range exts {
 		componentConfig := config.NewMapFromStringMap(value)
-		expandEnvConfig(componentConfig)
 
 		// Decode the key into type and fullName components.
-		id, err := config.NewIDFromString(key)
+		id, err := config.NewComponentIDFromString(key)
 		if err != nil {
 			return nil, errorInvalidTypeAndNameKey(extensionsKeyName, key, err)
 		}
@@ -263,7 +262,7 @@ func unmarshalService(rawService serviceSettings) (config.Service, error) {
 
 	ret.Extensions = make([]config.ComponentID, 0, len(rawService.Extensions))
 	for _, extIDStr := range rawService.Extensions {
-		id, err := config.NewIDFromString(extIDStr)
+		id, err := config.NewComponentIDFromString(extIDStr)
 		if err != nil {
 			return ret, err
 		}
@@ -301,10 +300,9 @@ func unmarshalReceivers(recvs map[string]map[string]interface{}, factories map[c
 	// Iterate over input map and create a config for each.
 	for key, value := range recvs {
 		componentConfig := config.NewMapFromStringMap(value)
-		expandEnvConfig(componentConfig)
 
 		// Decode the key into type and fullName components.
-		id, err := config.NewIDFromString(key)
+		id, err := config.NewComponentIDFromString(key)
 		if err != nil {
 			return nil, errorInvalidTypeAndNameKey(receiversKeyName, key, err)
 		}
@@ -338,10 +336,9 @@ func unmarshalExporters(exps map[string]map[string]interface{}, factories map[co
 	// Iterate over Exporters and create a config for each.
 	for key, value := range exps {
 		componentConfig := config.NewMapFromStringMap(value)
-		expandEnvConfig(componentConfig)
 
 		// Decode the key into type and fullName components.
-		id, err := config.NewIDFromString(key)
+		id, err := config.NewComponentIDFromString(key)
 		if err != nil {
 			return nil, errorInvalidTypeAndNameKey(exportersKeyName, key, err)
 		}
@@ -380,10 +377,9 @@ func unmarshalProcessors(procs map[string]map[string]interface{}, factories map[
 	// Iterate over processors and create a config for each.
 	for key, value := range procs {
 		componentConfig := config.NewMapFromStringMap(value)
-		expandEnvConfig(componentConfig)
 
 		// Decode the key into type and fullName components.
-		id, err := config.NewIDFromString(key)
+		id, err := config.NewComponentIDFromString(key)
 		if err != nil {
 			return nil, errorInvalidTypeAndNameKey(processorsKeyName, key, err)
 		}
@@ -422,7 +418,7 @@ func unmarshalPipelines(pipelinesConfig map[string]pipelineSettings) (config.Pip
 	// Iterate over input map and create a config for each.
 	for key, rawPipeline := range pipelinesConfig {
 		// Decode the key into type and name components.
-		id, err := config.NewIDFromString(key)
+		id, err := config.NewComponentIDFromString(key)
 		if err != nil {
 			return nil, errorInvalidTypeAndNameKey(pipelinesKeyName, key, err)
 		}
@@ -465,48 +461,13 @@ func unmarshalPipelines(pipelinesConfig map[string]pipelineSettings) (config.Pip
 func parseIDNames(pipelineID config.ComponentID, componentType string, names []string) ([]config.ComponentID, error) {
 	var ret []config.ComponentID
 	for _, idProcStr := range names {
-		idRecv, err := config.NewIDFromString(idProcStr)
+		idRecv, err := config.NewComponentIDFromString(idProcStr)
 		if err != nil {
 			return nil, fmt.Errorf("pipelines: config for %v contains invalid %s name %s : %w", pipelineID, componentType, idProcStr, err)
 		}
 		ret = append(ret, idRecv)
 	}
 	return ret, nil
-}
-
-// expandEnvConfig updates a config.Map with expanded values for all the values (simple, list or map value).
-// It does not expand the keys.
-func expandEnvConfig(v *config.Map) {
-	for _, k := range v.AllKeys() {
-		v.Set(k, expandStringValues(v.Get(k)))
-	}
-}
-
-func expandStringValues(value interface{}) interface{} {
-	switch v := value.(type) {
-	default:
-		return v
-	case string:
-		return expandEnv(v)
-	case []interface{}:
-		nslice := make([]interface{}, 0, len(v))
-		for _, vint := range v {
-			nslice = append(nslice, expandStringValues(vint))
-		}
-		return nslice
-	case map[string]interface{}:
-		nmap := make(map[interface{}]interface{}, len(v))
-		for k, vint := range v {
-			nmap[k] = expandStringValues(vint)
-		}
-		return nmap
-	case map[interface{}]interface{}:
-		nmap := make(map[interface{}]interface{}, len(v))
-		for k, vint := range v {
-			nmap[k] = expandStringValues(vint)
-		}
-		return nmap
-	}
 }
 
 // expandEnvLoadedConfig is a utility function that goes recursively through a config object