+++ exporter/exporterhelper/persistent_storage_batch.go
@@ -25,7 +25,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"go.opentelemetry.io/collector/extension/storage"
+	"go.opentelemetry.io/collector/extension/experimental/storage"
 )
 
 var errItemIndexArrInvalidDataType = errors.New("invalid data type, expected []itemIndex")